Release Python 3 version to PyPi
While the Python 3 code was already in it hadn't been uploaded to PyPi.
This commit adds full test support with tox, readme as RST to work on PyPi.

For tox the assumption is that the development is done using pyenv and
as such a .python-version for all the supported version is included.